BitDiff: Create to_size_t function and handle command line arguments
General outline complete. to_size_t converts a string to a size_t. This
is so we can convert the given size flag to an int that we use as the
size.

Main takes command line arguments. It should accept just a file but
throw a warning about not providing a size flag. It should not accept
just the size flag or no arguments at all. In either of those scenarios
it will throw an error and stop execution. Also, if the user provides
all the correct command line arguments but the file they provide fails
to open, an error is thrown and execution stops.